requirements.txt
pandas
office365-rest-python-client==2.5.14
